backend/services/conversation-service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv>=1.0.1
pydantic==2.5.3
sqlalchemy==2.0.25
pymysql==1.1.0
python-jose[cryptography]==3.3.0
pika==1.3.2
redis==5.0.1
httpx==0.27.0

